Fix to #26166 - Idempotent migration script generates problematic temporal table alter statement
Problem was that for idempotent migration scripts that convert non-temporal table to temporal we need to wrap the period switching statement in exec. Otherwise sql parser(?) complains that the command is invalid because period already exists. This is a false negative - the statement will not actually execute if this migration has already been applied. Exec fixes the problem since the statement inside it can't be parsed. This is how we deal with similar issues in other places in the code.

Fixes #26166