[6.0] Separate design time and runtime service for NamedConnectionStringResolver (#26205)
Fixes #26169

The issue here is that we started TryAdding design time services, which is good since it allows for overrides. However, it meant that for INamedConnectionStringResolver we were not replacing the runtime service, which then null-reffed at design time. Fix is to make them different services so one does not have to override the other.