Use correct value comparer for generated FK properties
Fixes #26116

Initial issue was that the converted type also required a comparer. On adding one it revealed that in this case a default comparer was being used for the dependent map in the state manager. Fix is to use the PK comparer in the same was as we use the PK converter.

Added more testing for similar cases including generated shadow FKs.