Fix to #26357 - [Q] SqlNullabilityProcessor.VisitSqlFunction, Nullability of SUM function
Marking SUM as non-nullable always. For the most part we were already doing the right thing during expand null semantics step (since it was evaluating nullability of the COALESCE(sum, 0) which is never null), but now we are bit more efficient.
Also, we can now apply some additional optimizations on SUM that required it to be non-nullable (e.g. pushing NOT operator inside)

Fixes #26357