Query: Avoid parethesis around binary operations when used with AND/OR
Sqlite/Postgre => AND - OR 2 most lowest priority operation in that order
SqlServer => Same as Sqlite except IN/LIKE which are below AND and same as OR
LIKE always puts parentheses around it. And IN evaluated left to right if first works. If 2nd then first one has to be same type as item (to cause `a OR b IN ...`) to evaluate OR first. That implies bool type and SqlServer cannot have bool value directly due to search condition so it always converts to `a = (1 as bit) OR b IN ...` which doesn't cause priority issue.