[6.0.1] Avoid infinite loop when finding principal in circular relationships
Fixes #26629

**Description**

We added code in 6.0 to find the corresponding principal key property of a foreign key property when looking for value comparers. However, we missed the case where the foreign key property is itself the principal key for another overlapping relationship. This can happen when two entity types act as principal and dependents of each other, with one having a composite foreign key containing the principal key of the other. In this case, the code goes into an infinite loop.

**Customer impact**

Stack overflow when a model of this form is used. This is not a super common model, but is also not extremely rare, so we expect many more people to hit this as the update to 6.0.

**How found**

Customer reported on 6.0.

**Regression**

Yes, from 5.0.

**Testing**

Added this model to our tests.

**Risk**

Low; keep track of the properties we have checked so we can avoid cycles. Also added a quirk mode.