Fix to #27356 - Collection_navigation_equal_to_null_for_subquery fails after merging from release/6.0
Improvement on the previous patch fix (#26744). We were not taking into the account scenario where collection projected from FirstOrDefault was compared to null. Collection can only be null if the parent entity is null, so we can safely apply the same optimization we do for anonymous types and required entities.

Fixes #27356