Fix to #28595 - Use partial updates for JSON
Currently, whenever we update/add/delete part of aggregate mapped to JSON column, we were replacing the entire structure.
Improvement is to use JSON_MODIFY which can alter just a portion of JSON structure, give a JSON path.
We analyze the entries that are being edited and for each JSON column we find the common denominator that needs to be replaced to accommodate all the requested changes

Note: Whenever we add/remove element from a collection we need to replace the entire collection in order to populate ordinal key values properly.
Also, modifying a single property is not yet supported - the smallest fragment that will be replaced is one that represents an entity.

Fixes #28595