Fix to #29279 - EF7 - GroupBy on Json column property + First/FirstOrDefault generates incorrect SQL
Problem was that in VisitChildren for JsonQueryExpression we were not passing key columns stored in _keyPropertyMap thru the visitor, so the aliases got out of sync.
Fix is to process these key columns also and then update the JsonQueryExpression with new map if any of the columns have been modified by the visitor.

Fixes #29279