Fix to #29355 - FromSqlRaw throws Exception when querying all objects that contain a certain string property in a json array column
Problem was that were not creating navigation bindings for JSON entities in the SelectExpression ctor which takes TableExpressionBase as argument. This would cause error in SharedTypeEntityExpandingExpressionVisitor which depends on those bindings to be present for entities mapped to JSON.

Fix is to use the same logic we use in the "main" SelectExpression ctor.

Fixes #29355