Ensure that facets specified in the store type name take precedence
Fixes #29600.

Using a store type name always overrides any other configuration of the type, since it is has ultimate fidelity of what the database supports. Previously, we were overriding facets set in the store type with those configured by data annotations of the model building API.

This change uses -1 to represent an explicitly configured "max" length, as opposed to no length configured, which is consistent with ADO.NET and EF6. Filed #29627 to make this first-class.