Fix to #30073 - JSON query regression in EF8
Problem was that when we were translating join, combining two SelectExpressions together, we were not remapping JSON column indexes stored in the inner select. This later result in shaper trying to read JSON column data from the wrong index and throwing the exception.

We already had a test that would have caught that, but it didn't return any data so the mismatch was never detected - fixed the data issue and added few more tests for good measure.

Fixes #30073