Fix to #30273 - EF7 generating incorrect SQL for the Concat/Union All, The same code was working fine for EF6.4.4
Problem was that during SelectExpression pruning for set operations we would try to prune both sources independently. However, if a source is Distinct we can't prune (that could affect the result). So the problematic case is where we prune one source but can't prune the second due to Distinct, and we end up with both sources having not matching projections.
Fix is to check if either source has Distinct before attempting to prune them.

Fixed #30273