Fix to #30244 - Json column fails to materialize only when default interceptor used (#30403)
* Fix to #30244 - Json column fails to materialize only when default interceptor used

Problem was that when we build materializer expression, ValueBufferTryReadValue call that extracts values of keys is typed with key's CLR type. This fails for json as we visit the materializer expression, because calls to ValueBufferTryReadValue are replaced with array accesses to object[] that we store key values in. We should be using ValueBufferTryReadValue<object> when trying to get value for keys, like we do everywhere else, so that after the visitation expression types are correct.

Fixes #30244

* Move testing to the interceptor tests, which are setup for singleton options