Stop skipping shadow skip navigation slots when creating the shadow values array
Fixes #30764

We were previously not allocating a slot for shadow skip navigations when values come from query, since they are always null. This is fine for the common case where the navigation values are at the end of the array. However, for a shadow navigation in an abstract base class, the slot may be in the middle of the array, which then means all our indexes are off by one (or more). To fix this, we now create slots for the shadow navs, even though they will always contain null.