Uses existing JSON primitive collection instead of creating a new one
Part of #30677

Builds on https://github.com/dotnet/efcore/pull/31272

This is the case where the CLR types contains, for example, an `ObservableCollection<int>`. In this case, we populate this existing collection, rather than creating a new instance.

Same still needs to be done for relational materialization of primitive collections.