Fix to #31597 - RC1: UnreachableException when loading with proxies and a JSON-mapped complex property
when processing entity materialization code, adapting it for JSON, we assume the entity type can be a result of BlockExpression (normal case) or a NewExpression (when the ctor is the only thing and we don't need to do any other assignments). However, for proxies, the block is wrapped around Convert node, since the instance generated of of Proxy type, but we expect the actual type there.

Fix is to handle this case and push the convert inside the block to normalize the expression we produce.

Fixes #31597