[release/8.0] Mitigation for stack overflow with compiled model on Mac
Fixes #31751

### Description

Using a large compiled model results in stack overflow on Mac due to excessive locals in the generated code. This change runs the compiled model initialization in a new thread with a bigger stack. This is a fix for 8 only--for 9 we will, factor the generated code differently.

### Customer impact

Stack overflow using the compiled model on a Mac.

### How found

Customer reported on RC1 and C.I. failure on Mac after adding a test with a big compiled model.

### Regression

Yes. The compiled model is bigger in EF8, which means that 6/7 models that worked will now throw.

### Testing

Updated tests, plus tests pass on Mac on the C.I. Also, reporting customer verified fix on his Mac.

### Risk

Low. Also quirked.