Fix to #31961 - Projection filtering on second-level reference could not be translated
Problem was that optimizing visitor was looking for a very specific pattern (member access over conditional), but in this scenario conditional is wrapped around Convert node.
Fix is to recognize this case also, strip convert around the conditional and apply it around non-null portion instead, before applying the member access over it.

Fixes #31961