Fix to #30478 - TemporalAll fails with JSON columns
We had a validation step preventing non AsOf temporal queries when we use navigation properties, as well as checking that all entities in the navigation chain are set as temporal. This should not apply to JSON entities,
since those always part of the entity.

Fix is to no longer perform the check for navigations mapped to JSON.

Fixes #30478