[release/8.0] Change error to warning for trigger defined on non-base type
Part of #31627

### Description

In EF8, we introduced validation that triggers are not defined on a non-base type in a TPH hierarchy. However, our guidance for configuring triggers on every type violates this. Therefore, we are updating the error to a warning, since having the trigger defined doesn't do much harm and so is likely to have been working in the past. We are also updating the guidance for configuring triggers on every type.

### Customer impact

Application that was working in EF7 will throw with EF8.

### How found

Customer reported on preview.

### Regression

Yes.

### Testing

Updated tests.

### Risk

Low.