Fix to #32023 - Case expressions may only be nested to level 10 when using SplitQuery() or ToList() inside projection. (#32094)
Problem was that our SqlExpression simplifying visitor, when visiting ShapedQuery would only look into QueryExpression part. This makes sense in general as QueryExpression is where the SqlExpression is stored. However, for split query scenarios, we also have SqlExpression in the ShaperExpression (wrapped by RelationalSplitCollectionShaperExpression), and all the optimizations were not run on that SqlExpression piece, which in that case caused too deep CASE WHEN ELSE structure (that we normally optimize into flat CASE with multiple WHENs.

Fix is to also visit ShaperExpression when performing post processing.

Fixes #32023