[release/8.0] Stop throwing for complex type properties when sensitive data logging is on
Fixes #32198

### Description

Property instances can now belong to complex types, which is a new feature in 8.0. However, logging of a change made to such a property incorrectly assumes the property must be on an entity type.

### Customer impact

Exception when using new feature in 8.0 when sensitive data logging is turned on, which is common.

### How found

Customer reported on 8.0 rc

### Regression

No; complex types are a new feature in 8.0

### Testing

Added tests.

### Risk

Very low; fixing an invalid cast.