Use current value in update pipeline when original value is not being tracked
Fixes #32168

The issue here is that when using changed/changing notifications, we don't store original values for non-key/index properties. However, in one path, the update pipeline is assuming that asking for original values is always safe since if there isn't one, then a current value is returned. However, this doesn't work if original values are _never_ saved for a given property. In that case, we need to explicitly ask for a current value instead.