Don't keep join table around because a ForeignKeyAttribute used to reference it
Fixes #27990

The issue here was a skip navigation with a `ForeignKeyAttribute`. At first I thought that this was not valid, but it turns out we decided that this configuration specified the name of the FK property in the join table.

However, this by-convention join type was then discarded and replaced with an explicitly configured join type. This normally removes the old join type from the model, but this doesn't happen here because the `ForeignKeyAttribute` resulted in the join type having a "data annotation" configuration source. This is not really correct; the FK attribute just says that the FK properties should have a given name, not anything about the type, so the entity type configuration source should not be changed here.

This allows the by-convention join type to be removed like it should, and hence we don't have two join tables mapped that differ only by case.