Stop throwing eagerly when looking up a ValueGenerator (#32930)
* Stop throwing eagerly when looking up a ValueGenerator

Fixes #32892

The issue was introduced by the fix for stable/unstable values. In that fix, we were missing that a value generator exists and influences whether the generated value is considered stable or not. However, some types, most notably bools, don't have a default generator. This is fine, it just means that there isn't a generator to use.

I removed the exception entirely because we never now throw if we can't find a generator--it's always okay. This is technically a breaking change, and people do create their own selectors, so we should probably doc it.

* Update to remove the breaking change and still throw for non-composite keys set for generation without any generator.

* Remove interface default implementation